tests: Added a base test for security and design
